MVP release
This is a really bare bones version that provides only the B2 buckets sizes and
file counts. The B2 API doe snot seem to have an efficient way of calculating
the sizes of buckets that doesn't involve iterating through the whole set of
objects. Consequently it takes over ten minutes for my B2 account to be counted.
I hope it doesn't run foul of any rate limits. As long at it doesn't the update
time is not a great problem as long as the update period is >> update time.